ELSIF-69: Fix review findings
Store `serialize_model` and `deserialize_model`  in one file.
`serialize_model` now excludes unused relations
 to reduce JSON size and exec time.
Sample JSON now contains predefined records instead of 20 random ones.
