+++ tests/README.md
@@ -6,7 +6,7 @@
 
 `mock_classes.py` - a module which contain base class with `httpx` request method, setup logic for each test suite and HTTP status codes
 
-`export_new_errata_rec.py` - a helper module for exporting NewErrataRecords from a DB to json file.
+`../scripts/serialize_new_errata_rec.py` - a helper module for exporting/importing NewErrataRecords from a DB to json file.
 
 ## How to run tests locally
 1. Adjust variables in `vars.env`
@@ -38,8 +38,9 @@
     Note: In order to export data you'll need DB with some data,\
     ideally the data must be as similar to production as possible
     ```bash
-    export SQLALCHEMY_URL="postgresql+psycopg2://<your-db-address>"
-    python3 tests/export_new_errata_rec.py
+    # set SQLALCHEMY_URL is it was not done on step 1
+    # export SQLALCHEMY_URL="postgresql+psycopg2://<your-db-address>"
+    python3 scripts/serialize_new_errata_rec.py
     ```
     Also, you'll need to update `tests/samples/test_oval.xml`.\
     For that uncomment block inside the test case in `tests/test_oval/test_oval_generation.py` file
